+++ VandasPage/frontend/src/components/utility/Dashboard.tsx
@@ -1,13 +1,8 @@
 import React from 'react';
 import { useState } from 'react';
 import Modal from './Modal';
-import './Dashboard.css';
-import DataContext from '../../context/dataContext';
-type Props = {
-  children: React.ReactElement;
-};
 
-const Dashboard = ({ children }: Props) => {
+const Dashboard = ({ children }) => {
   const [show, setShow] = useState(false);
 
   const showModal = () => {
@@ -17,20 +12,22 @@ const Dashboard = ({ children }: Props) => {
   const hideModal = () => {
     setShow(false);
   };
+  let updatedChildren = React.Children.map(children,
+    (child) => {
+        return React.cloneElement(child, { hideModal: hideModal });
+    })
+    return (
+      <>
+        <Modal show={show} handleClose={hideModal} >
+          {updatedChildren}
+        </Modal>
+        <button className='float-left mt-0 p-[0.12rem]' type="button" onClick={showModal}>
+          +
+        </button>
+      </>
+    );
+  }
 
-  let updatedChildren = React.Children.map(children, (child) => {
-    return React.cloneElement(child, { hideModal: hideModal });
-  });
-  return (
-    <>
-      <Modal show={show} handleClose={hideModal}>
-        {updatedChildren}
-      </Modal>
-      <button id="modalbutton" type="button" onClick={showModal}>
-        +
-      </button>
-    </>
-  );
-};
 
-export default Dashboard;
+
+export default Dashboard
